test: 당첨 번호에 대해서 테스트
